refactor: 레이아웃 컴포넌트 구조 변경
레이아웃과 관련된 컴포넌트 common -> layout 으로 이동
