+++ client/src/components/index.ts
@@ -1,1 +1,2 @@
-export { default as Layout } from './common/layout';
+export { default as Logo } from './common/logo';
+export { default as Layout } from './layout';
