+++ client/src/components/layout/footer.tsx
